cleaning(core): move PrismicSharedConfig
