+++ packages/core/src/node-utils/index.ts
@@ -2,6 +2,5 @@ export * as Framework from "./framework";
 export * from "./manifest";
 export * from "./paths";
 export * from "./pkg";
-export * from "./PrismicSharedConfig";
 export * as logs from "./cliLogs";
 export { default as Files } from "./files";
