+++ packages/core/src/prismic/index.ts
@@ -4,6 +4,7 @@ import { Repositories } from "../models/Repositories";
 
 export * as Communication from "./communication";
 export * as Endpoints from "./endpoints";
+export * from "./SharedConfig";
 
 export interface Core {
   cwd: string;
