+++ packages/slice-machine/server/src/api/services/getPrismicData.ts
@@ -1,18 +1,18 @@
 import { ok, err, Result } from "neverthrow";
 
 import PrismicData from "@lib/models/common/PrismicData";
-import { Utils, NodeUtils } from "@slicemachine/core";
+import { Utils, Prismic } from "@slicemachine/core";
 import ErrorWithStatus from "@lib/models/common/ErrorWithStatus";
 
 export default function getPrismicData(): Result<PrismicData, ErrorWithStatus> {
   try {
-    const prismicConfig = NodeUtils.PrismicSharedConfigManager.get();
+    const prismicConfig = Prismic.PrismicSharedConfigManager.get();
 
     const prismicData: PrismicData = {
       shortId: prismicConfig.shortId,
     };
 
-    if (prismicConfig.cookies === NodeUtils.DEFAULT_CONFIG.cookies) {
+    if (prismicConfig.cookies === Prismic.DEFAULT_CONFIG.cookies) {
       return ok(prismicData);
     }
